reduce realloc of tree traversal stack

+++ src/common.h
@@ -73,7 +73,7 @@ using cc::math::inverse;
 #define cosf(x) cc::math::cosf(x)
 #define sinf(x) cc::math::sinf(x)
 #define sincosf(x, s, c) cc::math::sincosf(x, s, c)
-#define powf(x, y) cc::math::powf(x, y)
+#define powf(x, y) cc::math::pow(x, y)
 #define sqrtf(x) cc::math::sqrtf(x)
 using cc::math::PI;
 using cc::array_size;
